# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        example: неверные данные
        type: string
    type: object
...
  model.Word:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      ex_jp:
        example:
        - 猫が好き
        items:
          type: string
        maxItems: 3
        type: array
      ex_ru:
        example:
        - Мне нравятся кошки
        items:
          type: string
        maxItems: 3
        type: array
      id:
        example: 1
        type: integer
      jp:
        example:
        - 猫
        items:
          type: string
        minItems: 1
        type: array
      kun:
        example:
        - ねこ
        items:
          type: string
        type: array
      "on":
        example:
        - ビョウ
        items:
          type: string
        type: array
      ru:
        example:
        - кот
        - кошка
        items:
          type: string
        minItems: 1
        type: array
      tags:
        example:
        - животные
        - JLPT N5
        items:
          type: string
        maxItems: 5
        type: array
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - jp
    - ru
    type: object
  model.WordCreateRequest:
    properties:
      ex_jp:
        example:
        - 猫が好き
        items:
          type: string
        maxItems: 3
        type: array
      ex_ru:
        example:
        - Мне нравятся кошки
        items:
          type: string
        maxItems: 3
        type: array
      jp:
        example:
        - 猫
        items:
          type: string
        minItems: 1
        type: array
      kun:
        example:
        - ねこ
        items:
          type: string
        type: array
      "on":
        example:
        - ビョウ
        items:
          type: string
        type: array
      ru:
        example:
        - кот
        - кошка
        items:
          type: string
        minItems: 1
        type: array
      tags:
        example:
        - животные
        items:
          type: string
        maxItems: 5
        type: array
    required:
    - jp
    - ru
    type: object
  model.WordResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/model.Word'
    type: object
  model.WordUpdateRequest:
    properties:
      ex_jp:
        example:
        - 猫が好き
        items:
          type: string
        maxItems: 3
        type: array
      ex_ru:
        example:
        - Мне нравятся кошки
        items:
          type: string
        maxItems: 3
        type: array
      jp:
        example:
        - 猫
        items:
          type: string
        minItems: 1
        type: array
      kun:
        example:
        - ねこ
        items:
          type: string
        type: array
      "on":
        example:
        - ビョウ
        items:
          type: string
        type: array
      ru:
        example:
        - кот
        - кошка
        items:
          type: string
        minItems: 1
        type: array
      tags:
        example:
        - животные
        items:
          type: string
        maxItems: 5
        type: array
    type: object
  model.WordsListResponse:
    properties:
      has_more:
        example: true
        type: boolean
      next_cursor:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      words:
        items:
          $ref: '#/definitions/model.Word'
        type: array
    type: object
  model.WordsListResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/model.WordsListResponse'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для личного русско-японского словаря
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JP-RU Dictionary API
  version: "1.0"
paths:
...
  /api/words:
    get:
      description: Возвращает список слов пользователя с курсорной пагинацией. Поддерживает
        фильтрацию и сортировку.
      parameters:
      - default: 20
        description: Лимит слов (макс. 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Курсор для пагинации (ID последнего загруженного слова)
        in: query
        minimum: 0
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список слов
          schema:
            $ref: '#/definitions/model.WordsListResponseWrapper'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список слов
      tags:
      - words
    post:
      consumes:
      - application/json
      description: Создает новое слово в словаре пользователя. Обязательны поля Jp
        (японский) и Ru (русский).
      parameters:
      - description: Данные слова
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.WordCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Слово создано
          schema:
            $ref: '#/definitions/model.WordResponseWrapper'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать слово
      tags:
      - words
  /api/words/{id}:
    delete:
      description: Удаляет слово по ID с проверкой владельца
      parameters:
      - description: ID слова
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Слово удалено
        "400":
          description: Неверный ID слова
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Слово не найдено
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить слово
      tags:
      - words
    get:
      description: Возвращает слово по ID с проверкой владельца
      parameters:
      - description: ID слова
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Слово
          schema:
            $ref: '#/definitions/model.WordResponseWrapper'
        "400":
          description: Неверный ID слова
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Слово не найдено
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить слово по ID
      tags:
      - words
    patch:
      consumes:
      - application/json
      description: |-
        Обновляет существующее слово (частичное обновление).
        Все поля опциональны: если поле не передано, оно сохраняет текущее значение.
        Передача пустого массива [] очищает поле.
        Передача null или отсутствие поля оставляет значение без изменений.
      parameters:
      - description: ID слова
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные (все поля опциональны)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.WordUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Слово обновлено
          schema:
            $ref: '#/definitions/model.WordResponseWrapper'
        "400":
          description: Неверные данные или пустой запрос
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Слово не найдено
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить слово
      tags:
      - words
...
  /api/words/search:
    get:
      description: Выполняет поиск слов по различным критериям
      parameters:
      - description: Поисковый запрос (подстрока в jp или ru)
        in: query
        name: q
        type: string
      - collectionFormat: csv
        description: Теги (точное совпадение)
        in: query
        items:
          type: string
        name: tags
        type: array
      - collectionFormat: csv
        description: Онъёми (точное совпадение)
        in: query
        items:
          type: string
        name: "on"
        type: array
      - collectionFormat: csv
        description: Кунъёми (точное совпадение)
        in: query
        items:
          type: string
        name: kun
        type: array
      - default: 20
        description: Лимит слов
        in: query
        name: limit
        type: integer
      - default: 0
        description: Курсор для пагинации
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Результаты поиска
          schema:
            $ref: '#/definitions/model.WordsListResponseWrapper'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Поиск слов
      tags:
      - words
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"